+++ CHANGELOG.md
@@ -17,6 +17,9 @@ Please avoid adding duplicate information across this changelog and JIRA/doc inp
     [`pfidpadapterid` parameter](https://docs.pingidentity.com/bundle/pingfederate-93/page/nfr1564003024683.html).
 - In `/v1/images` response, the `vulns` field `discoveredAt` is replaced by `firstSystemOccurrence` starting release 49.0. This field represents the first time the CVE was ever discovered in the system.
 - In `/v1/images` response, a new field `firstImageOccurrence` is added to `vulns` which represents the first time a CVE was discovered in respective image.
+- The default for the `--create-upgrader-sa` flag has changed to `true` in both the `roxctl sensor generate` and the
+  `roxctl sensor get-bundle` commands. In order to restore the old behavior, you need to explicitly specify
+  `--create-upgrader-sa=false`. 
 
 ## [48.0]
 - UI: Hovering over a namespace edge in the Network Graph will show the ports and protocols for it's connections (ROX-5228).